requirements.txt
djangorestframework == 3.11.2
djangorestframework-bulk == 0.2.1
djangorestframework-filters
django-admin-bootstrapped
django-bootstrap3 == 11.1.0
django-filter == 2.1.0
html5lib == 1.0.1
jdcal == 1.4.1
Markdown == 3.1.1
openpyxl == 2.6.2
lxml == 4.6.3
PyJWT == 1.7.1
requests == 2.23.0
python-dateutil == 2.8.1
pillow == 8.1.2
psutil == 5.7.3
https://github.com/frePPLe/django/tarball/frepple_6.0
psycopg2

# Additional packages for development environments
# black >= 19.3b0
# Sphinx
